removed deprecated methods and classes
