+++ pyincore/__init__.py
@@ -22,13 +22,7 @@ from pyincore.spaceservice import SpaceService
 from pyincore.utils.analysisutil import AnalysisUtil
 from pyincore.dataset import Dataset, InventoryDataset, DamageRatioDataset
 from pyincore.models.fragilitycurveset import FragilityCurveSet
-from pyincore.models.standardfragilitycurve import StandardFragilityCurve
-from pyincore.models.periodbuildingfragilitycurve import PeriodBuildingFragilityCurve
-from pyincore.models.periodstandardfragilitycurve import PeriodStandardFragilityCurve
-from pyincore.models.customexpressionfragilitycurve import CustomExpressionFragilityCurve
-from pyincore.models.conditionalstandardfragilitycurve import ConditionalStandardFragilityCurve
-from pyincore.models.parametricfragilitycurve import ParametricFragilityCurve
-from pyincore.models.fragilitycurverefactored import FragilityCurveRefactored
+from pyincore.models.fragilitycurve import FragilityCurve
 from pyincore.models.mappingset import MappingSet
 from pyincore.models.mapping import Mapping
 from pyincore.networkdata import NetworkData
